Add files via upload
Password managers are critical now, when a single person holds several accounts with credentials they would otherwise need to remember. If you’ve just learned about password managers or want to shop for one, it’s imperative to know what to look out for. There are dozens of options out there, and each password manager feature has its pros and cons.
But this Password manager help you store passwords, notes, and other personal credentials in a vault. This password manager can either store passwords locally, on your device, or in the Email.
you’ll be storing credentials for all your accounts, including important platforms like banks, security is of utmost importance. Luckily, password manager hold Cryptography encryption protocols to ensure your data is safe and secure from the bad guys.
Password manager is a software solution that supports the organization and the storing of passwords. Human-generated passwords are often algorithmically weak and prone to attacks. Password Manager is not only helpful for people, but is also helpful for different enterprises, such as small-sized businesses, medium-sized businesses, and large-sized businesses, as a tool for information and governance security. They are also used to secure, manage all passwords connected with their privileged EMAIL -accounts. Every Personnel can use it. With respective Requirement .There is no such limitation.
Features
Encrypted Using Ideaology of Cryptography that are highly Secure.
Encrypted Password are send through Email with key and description of password.
Generate QR code of the data.
Encrypted Password,Key & all info. are saved in form of TXT Files & QR Code.
It is more secure if key sent online through email server .so it is much more secure not much exposed to Internet world.
Under this project Cryptography is used ,It is a method of protecting information and communications through the use of codes, so that only those for whom the information is intended can read and process it.
In computer science, cryptography refers to secure information and communication techniques derived from mathematical concepts and a set of rule-based calculations called algorithms, to transform messages in ways that are hard to decipher. These deterministic algorithms are used for cryptographic key generation, digital signing, verification to protect data privacy, web browsing on the internet and confidential communications such as credit card transactions and email.